+++ R/sendrecv.R
@@ -64,10 +64,10 @@
 #'     corresponding to the vector sent should be used.
 #'
 #'     Mode 'next' sends serialised R objects, with native extensions enabled by
-#'     \code{\link{nextmode}}. This allows custom serialization and
-#'     unserialization functions to be registered for objects accessed via an
-#'     external pointer. When receiving, mode 'serial' should be used as 'next'
-#'     sends are fully compatible.
+#'     \code{\link{next_config}}. This configures custom serialization and
+#'     unserialization functions for external pointer type objects. When
+#'     receiving, mode 'serial' should be used as 'next' sends are fully
+#'     compatible.
 #'
 #' @seealso \code{\link{send_aio}} for asynchronous send.
 #' @examples